connected owner signup and dashboard
