+++ lib/firebase.js
@@ -8,6 +8,21 @@ const firebaseConfig = {
   projectId: process.env.FIREBASE_PROJECT_ID,
 };
 
+export const getOwnerType = async(email) => {
+  initFirebase();
+  console.log(email)
+  var snapshot = await firebase
+    .firestore()
+    .collection('users')
+    .where('email', '==', email)
+    .get()
+    .catch((err) => {
+      console.log(err);
+      return Promise.reject(err);
+    });
+  return snapshot.userType
+}
+
 const initFirebase = async () => {
   try {
     firebase.initializeApp(firebaseConfig);
@@ -49,7 +64,7 @@ export const signIn = async (email, password) => {
 
 export const signUp = async (userType, email, password,fullName, interests) => {
   initFirebase();
-
+  console.log(userType, email, password,fullName, interests)
   return firebase.auth().createUserWithEmailAndPassword(email, password)
   .then((userCredential) => {
     // Signed in 
@@ -201,8 +216,6 @@ async function getProjectDoc(projectId) {
   });
 
 }
-        
-
 
 export const getProjectsByDonor = async (userId) => {
   initFirebase();
